requirements.txt
absl-py==1.4.0
accelerate==0.18.0
adabelief-pytorch==0.2.1
aiohttp==3.8.3
aiosignal==1.3.1
alabaster==0.7.13
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
assertpy==1.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
autoflake==2.1.1
autopep8==2.0.1
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.2
bleach==6.0.0
blosc2==2.0.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
clip @ git+https://github.com/openai/CLIP.git@a9b1bf5920416aaeaec965c25dd9e8f98c864f16
cmake==3.25.0
colorama==0.4.6
comm==0.1.2
commonmark==0.9.1
contourpy==1.0.7
cryptography==39.0.1
cycler==0.11.0
Cython==0.29.33
datajoint==0.14.0
debugpy==1.6.6
decorator==5.1.1
decord==0.6.0
defusedxml==0.7.1
diffusers==0.14.0
dinov2 @ git+https://github.com/facebookresearch/dinov2.git@c3c2683a13cde94d4d99f523cf4170384b00c34c
distlib==0.3.6
docutils==0.19
einops==0.6.0
ema-pytorch==0.2.1
executing==1.2.0
fast-pytorch-kmeans==0.2.0.1
fastargs==1.2.0
fastjsonschema==2.16.2
ffcv==1.0.2
filelock==3.9.0
Flask==2.2.2
flatten-dict==0.4.2
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.1.0
ftfy==6.1.1
future==0.18.3
gitdb==4.0.10
GitPython==3.1.30
google-auth==2.16.0
google-auth-oauthlib==0.4.6
grpcio==1.51.1
h5py==3.8.0
hcp-utils==0.1.0
html5lib==1.1
huggingface-hub==0.13.3
idna==3.4
imageio==2.25.1
imagesize==1.4.1
importlib-metadata==6.0.0
importlib-resources==5.10.2
ipykernel==6.21.1
ipympl==0.9.3
ipython==8.9.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter_client==8.0.2
jupyter_core==5.2.0
jupyter_server==2.2.1
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
lightning-utilities==0.6.0.post0
lit==15.0.7
llvmlite==0.39.1
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
minio==7.1.13
mistune==2.0.5
mne==1.3.0
mpmath==1.2.1
msgpack==1.0.4
multidict==6.0.4
mypy-extensions==1.0.0
nbclassic==0.5.1
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==2.6.3
neuralpredictors==0.3.0
nibabel==5.0.0
nilearn==0.10.0
nnfabrik==0.2.1
notebook==6.5.2
notebook_shim==0.2.2
numba==0.56.4
numexpr==2.8.4
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
open-clip-torch==2.20.0
opencv-python==4.7.0.72
otumat==0.3.1
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pkgutil_resolve_name==1.3.10
platformdirs==2.6.2
pooch==1.6.0
positional-encodings==6.0.1
prometheus-client==0.16.0
prompt-toolkit==3.0.36
protobuf==3.20.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycortex==1.2.6
pycparser==2.21
pydot==1.4.2
pyflakes==3.0.1
Pygments==2.14.0
PyMySQL==1.0.2
pyparsing==3.0.9
pyre-extensions==0.0.29
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==0.21.1
python-json-logger==2.0.4
pytorch-lightning==1.9.4
pytorch-pfn-extras==0.6.5
pytorch-sphinx-theme==0.0.19
pytorch-triton==2.0.0+b8b470bc59
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.0
ray==2.2.0
recommonmark==0.7.1
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.2.1
scipy==1.10.0
seaborn==0.12.2
Send2Trash==1.8.0
shapely==2.0.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
snowballstemmer==2.2.0
soupsieve==2.3.2.post1
Sphinx==6.1.3
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
spikingjelly==0.0.0.0.12
stack-data==0.6.2
sympy==1.11.1
tables==3.8.0
tabulate==0.9.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
terminado==0.17.1
terminaltables==3.1.10
threadpoolctl==3.1.0
tifffile==2023.2.3
timm==0.9.2
tinycss2==1.2.1
tokenizers==0.13.2
tomli==2.0.1
torch-lazy @ git+https://github.com/simaki/torch-lazy@e3ce23b118bdf36a019c029a67bf5ec84f89a4d7
torch-tb-profiler==0.4.1
torchmetrics==0.11.1
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
transformers==4.27.3
triton==2.0.0
typing-inspect==0.8.0
typing_extensions==4.4.0
uri-template==1.2.0
urllib3==1.26.14
virtualenv==20.17.1
watchdog==2.2.1
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.2
wget==3.2
widgetsnbextension==4.0.5
xformers==0.0.20
yacs==0.1.8
yacs-stubgen==0.3.0
yarl==1.8.2
zipp==3.12.0
